Camera2: Maintain focus mode when using ttf
Add an option to maintain the focus mode when using
touch-to-focus

There's at least one family of camera HALs that gets "confused" when
the focus mode forcefully changes to auto after setting focus /
exposure area options.

Sticking with the current focus mode still results in effective use
of the focus region, so don't force a switch to auto. The default
behavior remains the original, this needs to be explicitly enabled
on affected devices

Change-Id: Icf70eb749cdfadd7c2739e41a9841681f2f3d561